+++ IPython/core/getipython.py
